Modernise code and dependencies
The previous version used strong encapsulation that made it difficult
to test individual functions in the code for generating nginx vhosts.
This version uses classes to allow individual methods to be tested.

This change also replaces many uses of callbacks with promises and
async/await, which required replacing the dependency on istanbul
with nyc. The restler dependency has been replaced with 'got',
as this has a promise-based API.

A script has been added that implements the same behaviour as
`grunt publish-redirects`. This will allow grunt to be removed
as a runtime dependency.